# Only set entry time when it will be logged