This pull request optimizes the Check scenario. It only sets the entry time after the decision has been made to log the entry or not. This improves performance in a disabled scenario by 63.55%.

```
benchmark                                           old ns/op     new ns/op     delta
BenchmarkDisabledAccumulatedContext/Zap.Check-4     20.3          7.40          -63.55%
```