@fgrosse 2019-01-01
We also noticed that recording the time in high throughput scenarios was a bottleneck, basically because we had a lot of debug messages. However I'm afraid you cannot merge this change without breaking compatibility because cores such as the sampling core rely on the time information in order to check if the log entry should be printed or not. Additionally there might be other hooked cores which work in a similar way.

What we ended up doing was to wrap the debug log calls in a function which only checks the level and aborts early if possible. Of course that means that we cannot use any sampling based logging there.