@pjvds 2019-01-02
Is there a reason why we don't remove the sampling state to the sampling core itself, or make it not time depending but just a percentage of the logs with a atomic counter? This would mean you only pay the overhead for sampling support if the sampling core is enabled.